fix: 백엔드 sql문 store_id의 타입을 int에서 long으로 변경

+++ backend/src/main/java/com/storerader/server/common/repository/sql/PriceRepositorySQL.java
@@ -78,7 +78,7 @@ public class PriceRepositorySQL {
                         )
                         """,
                 (rs, rowNum) -> new PriceItemDTO(
-                        rs.getInt("store_id"),
+                        rs.getLong("store_id"),
                         rs.getInt("price"),
                         rs.getString("inspect_day")
                 ),
